# Move data from Completion and ExpressionResult to context