--- a/Jint.Tests/Runtime/EngineLimitTests.cs
+++ b/Jint.Tests/Runtime/EngineLimitTests.cs
@@ -10,9 +10,9 @@ public class EngineLimitTests
     public void ShouldAllowReasonableCallStackDepth()
     {
 #if RELEASE
-        const int FunctionNestingCount = 750;
+        const int FunctionNestingCount = 810;
 #else
-        const int FunctionNestingCount = 420;
+        const int FunctionNestingCount = 460;
 #endif
 
         // generate call tree