--- a/Jint/Runtime/Completion.cs
+++ b/Jint/Runtime/Completion.cs
@@ -4,82 +4,63 @@ using Esprima;
 using Esprima.Ast;
 using Jint.Native;
 
-namespace Jint.Runtime
+namespace Jint.Runtime;
+
+public enum CompletionType : byte
 {
-    public enum CompletionType : byte
+    Normal = 0,
+    Return = 1,
+    Throw = 2,
+    Break,
+    Continue
+}
+
+/// <summary>
+/// https://tc39.es/ecma262/#sec-completion-record-specification-type
+/// </summary>
+[StructLayout(LayoutKind.Auto)]
+public readonly struct Completion
+{
+    private static readonly Node _emptyNode = new Identifier("");
+    private static readonly Completion _emptyCompletion = new(CompletionType.Normal, null!, _emptyNode);
+
+    internal readonly SyntaxElement _source;
+
+    public Completion(CompletionType type, JsValue value, SyntaxElement source)
     {
-        Normal = 0,
-        Return = 1,
-        Throw = 2,
-        Break,
-        Continue
+        Type = type;
+        Value = value;
+        _source = source;
+    }
+
+    public readonly CompletionType Type;
+    public readonly JsValue Value;
+    public ref readonly Location Location => ref _source.Location;
+
+    public static ref readonly Completion Empty() => ref _emptyCompletion;
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public JsValue GetValueOrDefault()
+    {
+        return Value ?? Undefined.Instance;
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public bool IsAbrupt()
+    {
+        return Type != CompletionType.Normal;
     }
 
     /// <summary>
-    /// https://tc39.es/ecma262/#sec-completion-record-specification-type
+    /// https://tc39.es/ecma262/#sec-updateempty
     /// </summary>
-    [StructLayout(LayoutKind.Auto)]
-    public readonly struct Completion
+    internal Completion UpdateEmpty(JsValue value)
     {
-        private static readonly Node _emptyNode = new Identifier("");
-        private static readonly Completion _emptyCompletion = new(CompletionType.Normal, null!, _emptyNode);
-
-        internal readonly SyntaxElement _source;
-
-        internal Completion(CompletionType type, JsValue value, string? target, SyntaxElement source)
+        if (Value is not null)
         {
-            Type = type;
-            Value = value;
-            Target = target;
-            _source = source;
+            return this;
         }
 
-        public Completion(CompletionType type, JsValue value, SyntaxElement source)
-        {
-            Type = type;
-            Value = value;
-            Target = null;
-            _source = source;
-        }
-
-        public Completion(CompletionType type, string target, SyntaxElement source)
-        {
-            Type = type;
-            Value = null!;
-            Target = target;
-            _source = source;
-        }
-
-        public readonly CompletionType Type;
-        public readonly JsValue Value;
-        public readonly string? Target;
-        public ref readonly Location Location => ref _source.Location;
-
-        public static ref readonly Completion Empty() => ref _emptyCompletion;
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public JsValue GetValueOrDefault()
-        {
-            return Value ?? Undefined.Instance;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public bool IsAbrupt()
-        {
-            return Type != CompletionType.Normal;
-        }
-
-        /// <summary>
-        /// https://tc39.es/ecma262/#sec-updateempty
-        /// </summary>
-        internal Completion UpdateEmpty(JsValue value)
-        {
-            if (Value is not null)
-            {
-                return this;
-            }
-
-            return new Completion(Type, value, Target, _source);
-        }
+        return new Completion(Type, value, _source);
     }
 }