--- a/Jint/Runtime/Environments/ExecutionContext.cs
+++ b/Jint/Runtime/Environments/ExecutionContext.cs
@@ -1,67 +1,68 @@
+using System.Runtime.InteropServices;
 using Jint.Native.Function;
 
-namespace Jint.Runtime.Environments
+namespace Jint.Runtime.Environments;
+
+[StructLayout(LayoutKind.Auto)]
+internal readonly struct ExecutionContext
 {
-    internal readonly struct ExecutionContext
+    internal ExecutionContext(
+        IScriptOrModule? scriptOrModule,
+        EnvironmentRecord lexicalEnvironment,
+        EnvironmentRecord variableEnvironment,
+        PrivateEnvironmentRecord? privateEnvironment,
+        Realm realm,
+        FunctionInstance? function = null)
     {
-        internal ExecutionContext(
-            IScriptOrModule? scriptOrModule,
-            EnvironmentRecord lexicalEnvironment,
-            EnvironmentRecord variableEnvironment,
-            PrivateEnvironmentRecord? privateEnvironment,
-            Realm realm,
-            FunctionInstance? function = null)
-        {
-            ScriptOrModule = scriptOrModule;
-            LexicalEnvironment = lexicalEnvironment;
-            VariableEnvironment = variableEnvironment;
-            PrivateEnvironment = privateEnvironment;
-            Realm = realm;
-            Function = function;
-        }
+        ScriptOrModule = scriptOrModule;
+        LexicalEnvironment = lexicalEnvironment;
+        VariableEnvironment = variableEnvironment;
+        PrivateEnvironment = privateEnvironment;
+        Realm = realm;
+        Function = function;
+    }
 
-        public readonly IScriptOrModule? ScriptOrModule;
-        public readonly EnvironmentRecord LexicalEnvironment;
-        public readonly EnvironmentRecord VariableEnvironment;
-        public readonly PrivateEnvironmentRecord? PrivateEnvironment;
-        public readonly Realm Realm;
-        public readonly FunctionInstance? Function;
+    public readonly IScriptOrModule? ScriptOrModule;
+    public readonly EnvironmentRecord LexicalEnvironment;
+    public readonly EnvironmentRecord VariableEnvironment;
+    public readonly PrivateEnvironmentRecord? PrivateEnvironment;
+    public readonly Realm Realm;
+    public readonly FunctionInstance? Function;
 
-        public ExecutionContext UpdateLexicalEnvironment(EnvironmentRecord lexicalEnvironment)
-        {
-            return new ExecutionContext(ScriptOrModule, lexicalEnvironment, VariableEnvironment, PrivateEnvironment, Realm, Function);
-        }
+    public ExecutionContext UpdateLexicalEnvironment(EnvironmentRecord lexicalEnvironment)
+    {
+        return new ExecutionContext(ScriptOrModule, lexicalEnvironment, VariableEnvironment, PrivateEnvironment, Realm, Function);
+    }
 
-        public ExecutionContext UpdateVariableEnvironment(EnvironmentRecord variableEnvironment)
-        {
-            return new ExecutionContext(ScriptOrModule, LexicalEnvironment, variableEnvironment, PrivateEnvironment, Realm, Function);
-        }
+    public ExecutionContext UpdateVariableEnvironment(EnvironmentRecord variableEnvironment)
+    {
+        return new ExecutionContext(ScriptOrModule, LexicalEnvironment, variableEnvironment, PrivateEnvironment, Realm, Function);
+    }
 
-        public ExecutionContext UpdatePrivateEnvironment(PrivateEnvironmentRecord? privateEnvironment)
-        {
-            return new ExecutionContext(ScriptOrModule, LexicalEnvironment, VariableEnvironment, privateEnvironment, Realm, Function);
-        }
+    public ExecutionContext UpdatePrivateEnvironment(PrivateEnvironmentRecord? privateEnvironment)
+    {
+        return new ExecutionContext(ScriptOrModule, LexicalEnvironment, VariableEnvironment, privateEnvironment, Realm, Function);
+    }
 
-        /// <summary>
-        /// https://tc39.es/ecma262/#sec-getthisenvironment
-        /// </summary>
-        internal EnvironmentRecord GetThisEnvironment()
+    /// <summary>
+    /// https://tc39.es/ecma262/#sec-getthisenvironment
+    /// </summary>
+    internal EnvironmentRecord GetThisEnvironment()
+    {
+        // The loop will always terminate because the list of environments always
+        // ends with the global environment which has a this binding.
+        var lex = LexicalEnvironment;
+        while (true)
         {
-            // The loop will always terminate because the list of environments always
-            // ends with the global environment which has a this binding.
-            var lex = LexicalEnvironment;
-            while (true)
+            if (lex != null)
             {
-                if (lex != null)
+                if (lex.HasThisBinding())
                 {
-                    if (lex.HasThisBinding())
-                    {
-                        return lex;
+                    return lex;
 
-                    }
-
-                    lex = lex._outerEnv;
                 }
+
+                lex = lex._outerEnv;
             }
         }
     }