--- a/Jint/Runtime/Interpreter/Statements/JintBreakStatement.cs
+++ b/Jint/Runtime/Interpreter/Statements/JintBreakStatement.cs
@@ -1,22 +1,19 @@
 using Esprima.Ast;
 
-namespace Jint.Runtime.Interpreter.Statements
+namespace Jint.Runtime.Interpreter.Statements;
+
+/// <summary>
+/// http://www.ecma-international.org/ecma-262/5.1/#sec-12.8
+/// </summary>
+internal sealed class JintBreakStatement : JintStatement<BreakStatement>
 {
-    /// <summary>
-    /// http://www.ecma-international.org/ecma-262/5.1/#sec-12.8
-    /// </summary>
-    internal sealed class JintBreakStatement : JintStatement<BreakStatement>
+    public JintBreakStatement(BreakStatement statement) : base(statement)
     {
-        private readonly string? _label;
+    }
 
-        public JintBreakStatement(BreakStatement statement) : base(statement)
-        {
-            _label = statement.Label?.Name;
-        }
-
-        protected override Completion ExecuteInternal(EvaluationContext context)
-        {
-            return new Completion(CompletionType.Break, null!, _label, _statement);
-        }
+    protected override Completion ExecuteInternal(EvaluationContext context)
+    {
+        context.Target = _statement.Label?.Name;
+        return new Completion(CompletionType.Break, null!, _statement);
     }
 }