--- a/Jint/Runtime/Interpreter/Statements/JintEmptyStatement.cs
+++ b/Jint/Runtime/Interpreter/Statements/JintEmptyStatement.cs
@@ -1,16 +1,15 @@
 using Esprima.Ast;
 
-namespace Jint.Runtime.Interpreter.Statements
-{
-    internal sealed class JintEmptyStatement : JintStatement<EmptyStatement>
-    {
-        public JintEmptyStatement(EmptyStatement statement) : base(statement)
-        {
-        }
+namespace Jint.Runtime.Interpreter.Statements;
 
-        protected override Completion ExecuteInternal(EvaluationContext context)
-        {
-            return new Completion(CompletionType.Normal, null!, null, _statement);
-        }
+internal sealed class JintEmptyStatement : JintStatement<EmptyStatement>
+{
+    public JintEmptyStatement(EmptyStatement statement) : base(statement)
+    {
+    }
+
+    protected override Completion ExecuteInternal(EvaluationContext context)
+    {
+        return new Completion(CompletionType.Normal, null!, _statement);
     }
 }