Small steps towards returning data using context instead of structs. `Target` is a prime example of returning possible jump instruction via completion which is not needed in 99% of the cases.

Tests seem to have quite the jitter, but show that we are moving to the right direction.

## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|dromaeo-3d-cube|27.893 ms|0.0420 ms|7166.38 KB|
| **New** |	|	| **26.489 ms (-5%)** | **0.0326 ms** | **7164.43 KB (0%)** |
| Old |Run|dromaeo-core-eval|7.049 ms|0.0172 ms|320.64 KB|
| **New** |	|	| **7.304 ms (+4%)** | **0.0526 ms** | **319.54 KB (0%)** |
| Old |Run|dromaeo-object-array|81.914 ms|0.3203 ms|106133.43 KB|
| **New** |	|	| **84.156 ms (+3%)** | **0.1687 ms** | **106134.6 KB (0%)** |
| Old |Run|droma(...)egexp [21]|315.073 ms|6.1587 ms|235918.41 KB|
| **New** |	|	| **307.709 ms (-2%)** | **2.6667 ms** | **234367.11 KB (-1%)** |
| Old |Run|droma(...)tring [21]|468.829 ms|34.3012 ms|1331640.95 KB|
| **New** |	|	| **438.789 ms (-6%)** | **31.2505 ms** | **1331443.9 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|75.731 ms|0.6149 ms|7789.08 KB|
| **New** |	|	| **72.038 ms (-5%)** | **0.1586 ms** | **7790.63 KB (0%)** |
| Old |Run|dromaeo-3d-cube|26.340 ms|0.0479 ms|6871.89 KB|
| **New** |	|	| **25.802 ms (-2%)** | **0.0327 ms** | **6869.94 KB (0%)** |
| Old |Run|dromaeo-core-eval|7.048 ms|0.0177 ms|308 KB|
| **New** |	|	| **7.690 ms (+9%)** | **0.0231 ms** | **306.9 KB (0%)** |
| Old |Run|dromaeo-object-array|84.917 ms|0.2181 ms|106094.9 KB|
| **New** |	|	| **81.887 ms (-4%)** | **0.2789 ms** | **106095.02 KB (0%)** |
| Old |Run|droma(...)egexp [21]|313.301 ms|6.1554 ms|234390.18 KB|
| **New** |	|	| **308.210 ms (-2%)** | **2.1133 ms** | **242594.01 KB (+4%)** |
| Old |Run|droma(...)tring [21]|433.911 ms|32.3667 ms|1331568.45 KB|
| **New** |	|	| **400.063 ms (-8%)** | **18.8510 ms** | **1331583.32 KB (0%)** |
| Old |Run|droma(...)ase64 [21]|75.744 ms|0.3382 ms|7701.19 KB|
| **New** |	|	| **69.859 ms (-8%)** | **0.1002 ms** | **7701.22 KB (0%)** |